remove ViewModel and ViewModelTests
